% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_die.R
\name{expand_dice}
\alias{expand_dice}
\title{Expands effects}
\usage{
expand_dice(x, lvl = 6, rate = 1)
}
\arguments{
\item{x}{numeric vector length >= 1.}

\item{lvl}{numeric scalar what MINIMUM result expands; default 6+.}

\item{rate}{numeric scalar number of actual successes on proc (default 1)?}
}
\value{
integer scalar number additional successes (see description).
}
\description{
Some effects spawn additional successes on certain rolls.
For example, on an attack roll of 6+, this attack scores 3 hits instead of 1.
I call this "expands" because it increases the number of successes without generating new rolls (explodes).
Expanding attacks have a rate of expansion -- how many new successes proc-d on flagged roll?
}
\details{
This function specifically returns the number of additionall successes, not including the initial success.
For example, if an attack grants 3 successes on a 6+, this function would return 2 -- the ADDITIONAL successes.
}
\examples{
x <- roll_dice(6)
expand_dice(x = x, lvl = 6, rate = 3)
}
