% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_die.R
\name{explode_dice}
\alias{explode_dice}
\title{Exploding dice}
\usage{
explode_dice(x, lvl = 6)
}
\arguments{
\item{x}{numeric vector length >= 1.}

\item{lvl}{numeric scalar what MINIMUM result explodes; default 6+.}
}
\value{
numeric vector of new rolls.
}
\description{
This function mimics Death to the False Emperor, Dakka Dakka Dakka, etc.
Some dice space new rolls on a 6+ (or even 5+).
This adds a new roll to the results for each explosion.
}
\examples{
x <- roll_dice(6)
# death to the false emperor
explode_die(x = x, lvl = 6)
}
