% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_die.R
\name{reroll_dice}
\alias{reroll_dice}
\title{Re-roll effects}
\usage{
reroll_dice(x, lvl = 1)
}
\arguments{
\item{x}{numeric vector length >= 1.}

\item{lvl}{numeric scalar what MAXIMUM result rerolls; default re-roll 1s.}
}
\value{
numeric vector of length == x.
}
\description{
Some effects cause some dice to be re-rolled e.g. SM Captain.
This function replaces a previous roll with a new value.
}
\examples{
x <- roll_dice(6)
# captain aura on to-hit
rr <- roll_dice(6)
reroll_dice(x = rr, lvl = 1)
}
