% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_die.R
\name{roll_dice}
\alias{roll_dice}
\title{Roll six-sided dice}
\usage{
roll_dice(n)
}
\arguments{
\item{n}{int scalar number of dice to roll.}
}
\value{
numeric vector of length >= 1.
}
\description{
Core function to roll a d6.
}
\examples{
roll_dice(1)
}
