% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_attack.R
\name{to_attack}
\alias{to_attack}
\title{How many successful hits from an attack?}
\usage{
to_attack(n, skill, reroll = NULL, explode = NULL, expand = NULL,
  trigger = NULL)
}
\arguments{
\item{n}{integer scalar number of attacks}

\item{skill}{integer scalar minimum successful roll for relevant attack skill (e.g. BS 3+ would be skill = 3).}

\item{reroll}{integer scalar what MAXIMUM result rerolls (default NULL).}

\item{explode}{integer scalar what MINIMUM result explodes (default NULL).}

\item{expand}{named list of arguments for \code{\link{expand_dice}} (default NULL).}

\item{trigger}{integer vector length >= 1 which values trigger special effects (default NULL).}
}
\value{
integer scalar number of successful hits UNLESS !is.null(trigger) then named list with number of successes (success) and triggers (trigger).
}
\description{
When a unit attacks another unit, dice are rolled against the ballistic or weapon skill of the attacker.
Successful attacks have the opportunity to wound the defender.
}
\examples{
# a single SM tactical attacks.
to_attack(n = 1, skill = 3)

# a single SM tactical attacks while within 6" of a Captain.
to_attack(n = 1, skill = 3, reroll = 1)

# a single CSM attacks with Death to the False Emperor.
to_attack(n = 1, skill = 3, explode = 6)

# with expanding attacks
to_attack(n = 1, skill = 3, expand = list(lvl = 6, rate = 3))

# a single SM tactical attacks with an overcharged plasma
to_attack(n = 1, skill = 3, trigger = 1)
}
