% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_save.R
\name{to_fnp}
\alias{to_fnp}
\title{Feel No Pain}
\usage{
to_fnp(n, fnp)
}
\arguments{
\item{n}{integer scalar number of un-saved wounds}

\item{fnp}{integer scalar minimum to ignore wound (e.g. fnp = 5 for 5+ Disgustingly Resilient)}
}
\value{
integer scalar number of failed FNPs
}
\description{
After a failed save, some rules allow an additional save-after-the-save to prevent wounds individually.
This is sometimes refered as a shrug.
This function stands in place of e.g. Disgustingly Resilient.
8th does not have universal rules, but this function uses the older name for convenience.
}
\examples{
# CSM plague marine attempts to shrug a wound.
to_fnp(1, 5)
}
