% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_save.R
\name{to_save}
\alias{to_save}
\title{Save against wound roll}
\usage{
to_save(n, sv, ap = 0, invul = FALSE)
}
\arguments{
\item{n}{integer scalar number of wounds}

\item{sv}{integer scalar minimum save value (e.g. SM marine has SV 3+).}

\item{ap}{integer scalar armor piercing value of wound; negative and positive values accepted e.g. AP -3 and AP 3 are equivalent (default 0).}

\item{invul}{logical scalar is the save an invulnerability save? default FALSE.}
}
\value{
integer scalar number of failed saves
}
\description{
After a successful wound, the defender gets an opportunity to save and prevent that wound.
}
\examples{
# SM tactical marine saves against a boltgun hit.
to_save(n = 1, sv = 3, ap = 0)
}
