% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_wound.R
\name{to_wound}
\alias{to_wound}
\title{How many hits wound the target?}
\usage{
to_wound(n, str, tgh, reroll = NULL, explode = NULL, expand = NULL)
}
\arguments{
\item{n}{integer scalar number of hits.}

\item{str}{integer scalar strength of the attack.}

\item{tgh}{integer scalar toughness of the target.}

\item{reroll}{integer scalar what MAXIMUM result rerolls (default NULL).}

\item{explode}{integer scalar what MINIMUM result explodes (default NULL).}

\item{expand}{named list of arguments for \code{\link{expand_dice}} (default NULL).}
}
\value{
integer scalar number of wounding hits.
}
\description{
After a successful hit, the attacked determines how many wounds are dealt.
}
\examples{
# hit a SM tactical with a boltgun -- how many wounds?
to_wound(n = 1, str = 4, tgh = 4)

# hit a SM tactical with a boltgun within 6" of a SM Lt -- how many wounds?
to_wound(n = 1, str = 4, tgh = 4, reroll = 1)
}
