% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_die.R
\name{trigger_dice}
\alias{trigger_dice}
\title{Trigger dice}
\usage{
trigger_dice(x, trigger)
}
\arguments{
\item{x}{numeric vector length >= 1 of dice rolls.}

\item{trigger}{numeric vector of length >= 1. What EXACT values are special?}
}
\value{
numeric scalar number of triggers proc-d.
}
\description{
This function identifies how many triggers are proc-d by a roll.
A trigger is a specific value (or values) which are associated with special events.
For example, an overcharged plasma has a trigger on an attack roll of 1.
An overcharged plasma with a -1 to hit would trigger on attack rolls of 1 AND 2.
}
\examples{
# trigger on a 1 (e.g. overcharged plasma)
rr <- roll_dice(1)
trigger_dice(x = rr, trigger = 1)
}
